refactor: split views/utils/selectors into focused modules
Converts the 638-line selectors.ts into a views/utils/selectors/ directory
layered as base / fleet / ship / equip / map-sortie / aggregate /
const-derived, with index.ts re-exporting the whole surface. The module
specifier 'views/utils/selectors' resolves to the directory unchanged for
the 21 in-repo importers and for third-party plugins.

Pure move with two structural notes:
- fleetStateSelectorFactory lives in aggregate.ts (not fleet.ts): it
  consumes fleetShipsDataSelectorFactory, which would otherwise create a
  circular import; the original file relied on lazy evaluation for this
  forward reference.
- getDeckState moves next to its only consumer (aggregate.ts); a few
  previously module-private helpers (arrayResultWrapper, getMapData,
  getMapHp, shipBaseDataSelectorFactory) are now exported for cross-file
  use, which additively widens the barrel surface.
