+++ views/utils/selectors/index.ts
@@ -1,0 +1,11 @@
+// Redux selectors, split by domain. Layering (each file only imports from
+// earlier ones): base → fleet → ship → equip → map-sortie → aggregate →
+// const-derived. This barrel re-exports the whole surface so the module
+// specifier 'views/utils/selectors' keeps working for poi and for plugins.
+export * from './base'
+export * from './fleet'
+export * from './ship'
+export * from './equip'
+export * from './map-sortie'
+export * from './aggregate'
+export * from './const-derived'
